MAINT: Move initialization strategy to separate file
